requirements.txt
streamlit
pandas
azure-storage-blob
azure-identity
python-dotenv
